rename former owner_rpc_s to owner_rpc
