+++ api/src/lib.rs
@@ -41,18 +41,18 @@ mod foreign;
 mod foreign_rpc;
 
 mod owner;
-mod owner_rpc_s;
+mod owner_rpc;
 
 mod types;
 
 pub use crate::foreign::{Foreign, ForeignCheckMiddleware, ForeignCheckMiddlewareFn};
 pub use crate::foreign_rpc::ForeignRpc;
 pub use crate::owner::Owner;
-pub use crate::owner_rpc_s::OwnerRpcS;
+pub use crate::owner_rpc::OwnerRpc;
 
 pub use crate::foreign_rpc::foreign_rpc as foreign_rpc_client;
 pub use crate::foreign_rpc::run_doctest_foreign;
-pub use crate::owner_rpc_s::run_doctest_owner;
+pub use crate::owner_rpc::run_doctest_owner;
 
 pub use types::{
 	ECDHPubkey, EncryptedRequest, EncryptedResponse, EncryptionErrorResponse, JsonId, PubAddress,
