+++ api/src/owner_rpc.rs
@@ -15,27 +15,34 @@
 //! JSON-RPC Stub generation for the Owner API
 use uuid::Uuid;
 
+use crate::config::{TorConfig, WalletConfig};
 use crate::core::core::Transaction;
+use crate::core::global;
 use crate::keychain::{Identifier, Keychain};
 use crate::libwallet::slate_versions::v3::TransactionV3;
 use crate::libwallet::{
 	AcctPathMapping, ErrorKind, InitTxArgs, IssueInvoiceTxArgs, NodeClient, NodeHeightResult,
-	OutputCommitMapping, Slate, SlateVersion, TxLogEntry, VersionedSlate, WalletInfo,
-	WalletLCProvider,
+	OutputCommitMapping, PaymentProof, Slate, SlateVersion, StatusMessage, TxLogEntry,
+	VersionedSlate, WalletInfo, WalletLCProvider,
 };
-use crate::util::Mutex;
-use crate::{Owner, OwnerRpcS};
+use crate::util::logger::LoggingConfig;
+use crate::util::secp::key::{PublicKey, SecretKey};
+use crate::util::{static_secp_instance, Mutex, ZeroingString};
+use crate::{ECDHPubkey, Owner, PubAddress, Token};
 use easy_jsonrpc_mw;
 use grin_wallet_util::OnionV3Address;
+use rand::thread_rng;
 use std::convert::TryFrom;
 use std::sync::Arc;
+use std::time::Duration;
 
 /// Public definition used to generate Owner jsonrpc api.
-/// * When running `grin-wallet owner_api` with defaults, the V2 api is available at
-/// `localhost:3420/v2/owner`
-/// * The endpoint only supports POST operations, with the json-rpc request as the body
+/// Secure version containing wallet lifecycle functions. All calls to this API must be encrypted.
+/// See [`init_secure_api`](#tymethod.init_secure_api) for details of secret derivation
+/// and encryption.
+
 #[easy_jsonrpc_mw::rpc]
-pub trait OwnerRpc: Sync + Send {
+pub trait OwnerRpc {
 	/**
 	Networked version of [Owner::accounts](struct.Owner.html#method.accounts).
 
@@ -47,7 +54,9 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "accounts",
-		"params": [],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000"
+		},
 		"id": 1
 	}
 	# "#
@@ -66,14 +75,10 @@ pub trait OwnerRpc: Sync + Send {
 		"id": 1
 	}
 	# "#
-	# , false, 4, false, false, false, false);
+	# , true, 4, false, false, false, false);
 	```
 	*/
-	#[deprecated(
-		since = "3.0.0",
-		note = "The V2 Owner API (OwnerRpc) will be removed in grin-wallet 4.0.0. Please migrate to the V3 (OwnerRpcS) API as soon as possible."
-	)]
-	fn accounts(&self) -> Result<Vec<AcctPathMapping>, ErrorKind>;
+	fn accounts(&self, token: Token) -> Result<Vec<AcctPathMapping>, ErrorKind>;
 
 	/**
 	Networked version of [Owner::create_account_path](struct.Owner.html#method.create_account_path).
@@ -86,7 +91,10 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "create_account_path",
-		"params": ["account1"],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"label": "account1"
+		},
 		"id": 1
 	}
 	# "#
@@ -100,10 +108,10 @@ pub trait OwnerRpc: Sync + Send {
 		"id": 1
 	}
 	# "#
-	# ,false, 4, false, false, false, false);
+	# ,true, 4, false, false, false, false);
 	```
 	 */
-	fn create_account_path(&self, label: &String) -> Result<Identifier, ErrorKind>;
+	fn create_account_path(&self, token: Token, label: &String) -> Result<Identifier, ErrorKind>;
 
 	/**
 	Networked version of [Owner::set_active_account](struct.Owner.html#method.set_active_account).
@@ -116,7 +124,10 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "set_active_account",
-		"params": ["default"],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"label": "default"
+		},
 		"id": 1
 	}
 	# "#
@@ -130,10 +141,10 @@ pub trait OwnerRpc: Sync + Send {
 		"id": 1
 	}
 	# "#
-	# , false, 4, false, false, false, false);
+	# , true, 4, false, false, false, false);
 	```
 	 */
-	fn set_active_account(&self, label: &String) -> Result<(), ErrorKind>;
+	fn set_active_account(&self, token: Token, label: &String) -> Result<(), ErrorKind>;
 
 	/**
 	Networked version of [Owner::retrieve_outputs](struct.Owner.html#method.retrieve_outputs).
@@ -146,7 +157,12 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "retrieve_outputs",
-		"params": [false, true, null],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"include_spent": false,
+			"refresh_from_node": true,
+			"tx_id": null
+		},
 		"id": 1
 	}
 	# "#
@@ -196,11 +212,12 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# , false, 2, false, false, false, false);
+	# , true, 2, false, false, false, false);
 	```
 	*/
 	fn retrieve_outputs(
 		&self,
+		token: Token,
 		include_spent: bool,
 		refresh_from_node: bool,
 		tx_id: Option<u32>,
@@ -217,7 +234,12 @@ pub trait OwnerRpc: Sync + Send {
 		{
 			"jsonrpc": "2.0",
 			"method": "retrieve_txs",
-			"params": [true, null, null],
+			"params": {
+				"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+				"refresh_from_node": true,
+				"tx_id": null,
+				"tx_slate_id": null
+			},
 			"id": 1
 		}
 		# "#
@@ -266,8 +288,8 @@ pub trait OwnerRpc: Sync + Send {
 			  "parent_key_id": "0200000000000000000000000000000000",
 			  "stored_tx": null,
 			  "ttl_cutoff_height": null,
-			  "tx_slate_id": null,
 			  "payment_proof": null,
+			  "tx_slate_id": null,
 			  "tx_type": "ConfirmedCoinbase"
 			}
 		  ]
@@ -275,12 +297,13 @@ pub trait OwnerRpc: Sync + Send {
 	  }
 	}
 	# "#
-	# , false, 2, false, false, false, false);
+	# , true, 2, false, false, false, false);
 	```
 	*/
 
 	fn retrieve_txs(
 		&self,
+		token: Token,
 		refresh_from_node: bool,
 		tx_id: Option<u32>,
 		tx_slate_id: Option<Uuid>,
@@ -295,7 +318,11 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "retrieve_summary_info",
-		"params": [true, 1],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"refresh_from_node": true,
+			"minimum_confirmations": 1
+		},
 		"id": 1
 	}
 	# "#
@@ -321,12 +348,13 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# ,false, 4, false, false, false, false);
+	# ,true, 4, false, false, false, false);
 	```
 	 */
 
 	fn retrieve_summary_info(
 		&self,
+		token: Token,
 		refresh_from_node: bool,
 		minimum_confirmations: u64,
 	) -> Result<(bool, WalletInfo), ErrorKind>;
@@ -341,6 +369,7 @@ pub trait OwnerRpc: Sync + Send {
 			"jsonrpc": "2.0",
 			"method": "init_send_tx",
 			"params": {
+				"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
 				"args": {
 					"src_acct_name": null,
 					"amount": "6000000000",
@@ -350,7 +379,7 @@ pub trait OwnerRpc: Sync + Send {
 					"selection_strategy_is_use_all": true,
 					"message": "my message",
 					"target_slate_version": null,
-					"payment_proof_recipient_address": null,
+					"payment_proof_recipient_address": "pa7wkkdgs5bkteha7lykl7ff2wztgdrxxo442xdcq2lnaphe5aidd4id",
 					"ttl_blocks": null,
 					"send_args": null
 				}
@@ -371,8 +400,12 @@ pub trait OwnerRpc: Sync + Send {
 		  "id": "0436430c-2b02-624c-2032-570501212b00",
 		  "lock_height": "0",
 			"ttl_cutoff_height": null,
-			"payment_proof": null,
 		  "num_participants": 2,
+			"payment_proof": {
+		"receiver_address": "783f6528669742a990e0faf0a5fca5d5b3330e37bbb9cd5c628696d03ce4e810",
+		"receiver_signature": null,
+		"sender_address": "32cdd63928854f8b2628b1dce4626ddcdf35d56cb7cfdf7d64cca5822b78d4d3"
+	  },
 		  "participant_data": [
 			{
 			  "id": "0",
@@ -419,11 +452,11 @@ pub trait OwnerRpc: Sync + Send {
 	  }
 	}
 		# "#
-		# ,false, 4, false, false, false, false);
+		# ,true, 4, false, false, false, false);
 	```
 	*/
 
-	fn init_send_tx(&self, args: InitTxArgs) -> Result<VersionedSlate, ErrorKind>;
+	fn init_send_tx(&self, token: Token, args: InitTxArgs) -> Result<VersionedSlate, ErrorKind>;
 
 	/**
 		Networked version of [Owner::issue_invoice_tx](struct.Owner.html#method.issue_invoice_tx).
@@ -435,6 +468,7 @@ pub trait OwnerRpc: Sync + Send {
 			"jsonrpc": "2.0",
 			"method": "issue_invoice_tx",
 			"params": {
+				"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
 				"args": {
 					"amount": "6000000000",
 					"message": "Please give me your grins",
@@ -458,8 +492,8 @@ pub trait OwnerRpc: Sync + Send {
 					"id": "0436430c-2b02-624c-2032-570501212b00",
 					"lock_height": "0",
 					"ttl_cutoff_height": null,
-					"payment_proof": null,
 					"num_participants": 2,
+					"payment_proof": null,
 					"participant_data": [
 						{
 							"id": "1",
@@ -501,11 +535,15 @@ pub trait OwnerRpc: Sync + Send {
 			}
 		}
 		# "#
-		# ,false, 4, false, false, false, false);
+		# ,true, 4, false, false, false, false);
 	```
 	*/
 
-	fn issue_invoice_tx(&self, args: IssueInvoiceTxArgs) -> Result<VersionedSlate, ErrorKind>;
+	fn issue_invoice_tx(
+		&self,
+		token: Token,
+		args: IssueInvoiceTxArgs,
+	) -> Result<VersionedSlate, ErrorKind>;
 
 	/**
 		 Networked version of [Owner::process_invoice_tx](struct.Owner.html#method.process_invoice_tx).
@@ -516,16 +554,17 @@ pub trait OwnerRpc: Sync + Send {
 		{
 			"jsonrpc": "2.0",
 			"method": "process_invoice_tx",
-			"params": [
-				{
+			"params": {
+				"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+				"slate": {
 					"amount": "6000000000",
 					"fee": "0",
 					"height": "4",
 					"id": "0436430c-2b02-624c-2032-570501212b00",
 					"lock_height": "0",
 					"ttl_cutoff_height": null,
-					"payment_proof": null,
 					"num_participants": 2,
+					"payment_proof": null,
 					"participant_data": [
 						{
 							"id": "1",
@@ -564,7 +603,7 @@ pub trait OwnerRpc: Sync + Send {
 						"block_header_version": 2
 					}
 				},
-				{
+				"args": {
 					"src_acct_name": null,
 					"amount": "0",
 					"minimum_confirmations": 2,
@@ -577,7 +616,7 @@ pub trait OwnerRpc: Sync + Send {
 					"ttl_blocks": null,
 					"send_args": null
 				}
-			],
+			},
 			"id": 1
 		}
 		# "#
@@ -594,8 +633,8 @@ pub trait OwnerRpc: Sync + Send {
 				"id": "0436430c-2b02-624c-2032-570501212b00",
 				"lock_height": "0",
 				"ttl_cutoff_height": null,
-				"payment_proof": null,
 				"num_participants": 2,
+				"payment_proof": null,
 				"participant_data": [
 					{
 						"id": "1",
@@ -655,12 +694,13 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# ,false, 4, false, false, false, false);
+	# ,true, 4, false, false, false, false);
 	```
 	*/
 
 	fn process_invoice_tx(
 		&self,
+		token: Token,
 		slate: VersionedSlate,
 		args: InitTxArgs,
 	) -> Result<VersionedSlate, ErrorKind>;
@@ -675,15 +715,17 @@ pub trait OwnerRpc: Sync + Send {
 		"jsonrpc": "2.0",
 		"method": "tx_lock_outputs",
 		"id": 1,
-		"params": [ {
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"slate": {
 				"amount": "6000000000",
 				"fee": "8000000",
 				"height": "4",
 				"id": "0436430c-2b02-624c-2032-570501212b00",
 				"lock_height": "4",
 				"ttl_cutoff_height": null,
-				"payment_proof": null,
 				"num_participants": 2,
+				"payment_proof": null,
 				"participant_data": [
 				{
 					"id": "0",
@@ -727,8 +769,8 @@ pub trait OwnerRpc: Sync + Send {
 					"block_header_version": 2
 				}
 			},
-			0
-		]
+			"participant_id": 0
+		}
 	}
 	# "#
 	# ,
@@ -741,12 +783,13 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# ,false, 5 ,true, false, false, false);
+	# ,true, 5 ,true, false, false, false);
 
 	```
 	 */
 	fn tx_lock_outputs(
 		&self,
+		token: Token,
 		slate: VersionedSlate,
 		participant_id: usize,
 	) -> Result<(), ErrorKind>;
@@ -761,77 +804,78 @@ pub trait OwnerRpc: Sync + Send {
 		"jsonrpc": "2.0",
 		"method": "finalize_tx",
 		"id": 1,
-		"params": [
-		{
-			"version_info": {
-				"version": 3,
-				"orig_version": 3,
-				"block_header_version": 2
-			},
-			"num_participants": 2,
-			"id": "0436430c-2b02-624c-2032-570501212b00",
-			"ttl_cutoff_height": null,
-			"payment_proof": null,
-			"tx": {
-				"offset": "d202964900000000d302964900000000d402964900000000d502964900000000",
-				"body": {
-					"inputs": [
-						{
-							"features": "Coinbase",
-							"commit": "087df32304c5d4ae8b2af0bc31e700019d722910ef87dd4eec3197b80b207e3045"
-						},
-						{
-							"features": "Coinbase",
-							"commit": "08e1da9e6dc4d6e808a718b2f110a991dd775d65ce5ae408a4e1f002a4961aa9e7"
-						}
-					],
-					"outputs": [
-						{
-							"features": "Plain",
-							"commit": "099b48cfb1f80a2347dc89818449e68e76a3c6817a532a8e9ef2b4a5ccf4363850",
-							"proof": "7ebcd2ed9bf5fb29854033ba3d0e720613bdf7dfacc586d2f6084c1cde0a2b72e955d4ce625916701dc7c347132f40d0f102a34e801d745ee54b49b765d08aae0bb801c60403e57cafade3b4b174e795b633ab9e402b5b1b6e1243fd10bbcf9368a75cb6a6c375c7bdf02da9e03b7f210df45d942e6fba2729cd512a372e6ed91a1b5c9c22831febea843e3f85adcf198f39ac9f7b73b70c60bfb474aa69878ea8d1d32fef30166b59caacaec3fd024de29a90f1587e08d2c36b3d5c560cabf658e212e0a40a4129b3e5c35557058def5551f4eb395759597ba808b3c34eac3bfb9716e4480d7931c5789c538463ec75be0eb807c894047fda6cbcd22682d3c6d3823cb330f090a2099e3510a3706b57d46c95224394d7f1c0a20d99cc314b8f1d9d02668e2e435f62e1194de0be6a1f50f72ed777ed51c8819f527a94918d1aa8df6461e98ed4c2b18210de50fbcf8c3df210bfe326d41f1dc0ad748cb0320ae28401c85ab4f7dcb99d88a052e95dc85b76d22b36cabd60e06ab84bb7e4ddfdab9c9730c8a986583237ed1ecbb323ee8e79b8cadca4b438b7c09531670b471dda6a2eb3e747916c88ce7d9d8e1b7f61660eeb9e5a13c60e4dfe89d1177d81d6f6570fda85158e646a15f1e8b9e977494dc19a339aab2e0e478670d80092d6ba37646e60714ef64eb4a3d37fe15f8f38b59114af34b235489eed3f69b7781c5fe496eb43ffe245c14bd740f745844a38cf0d904347aaa2b64f51add18822dac009d8b63fa3e4c9b1fa72187f9a4acba1ab315daa1b04c9a41f3be846ac420b37990e6c947a16cc9d5c0671b292bf77d7d8b8974d2ad3afae95ba7772c37432840f53a007f31e0195f3abdf100c4477723cc6c6d5da14894a73dfac342833731036487488fdade7b9d556c06f26173b6b67598d3769447ce2828d71dd45ac5af436c6b0"
-						},
-						{
-							"features": "Plain",
-							"commit": "0812276cc788e6870612296d926cba9f0e7b9810670710b5a6e6f1ba006d395774",
-							"proof": "dcff6175390c602bfa92c2ffd1a9b2d84dcc9ea941f6f317bdd0f875244ef23e696fd17c71df79760ce5ce1a96aab1d15dd057358dc835e972febeb86d50ccec0dad7cfe0246d742eb753cf7b88c045d15bc7123f8cf7155647ccf663fca92a83c9a65d0ed756ea7ebffd2cac90c380a102ed9caaa355d175ed0bf58d3ac2f5e909d6c447dfc6b605e04925c2b17c33ebd1908c965a5541ea5d2ed45a0958e6402f89d7a56df1992e036d836e74017e73ccad5cb3a82b8e139e309792a31b15f3ffd72ed033253428c156c2b9799458a25c1da65b719780a22de7fe7f437ae2fccd22cf7ea357ab5aa66a5ef7d71fb0dc64aa0b5761f68278062bb39bb296c787e4cabc5e2a2933a416ce1c9a9696160386449c437e9120f7bb26e5b0e74d1f2e7d5bcd7aafb2a92b87d1548f1f911fb06af7bd6cc13cee29f7c9cb79021aed18186272af0e9d189ec107c81a8a3aeb4782b0d950e4881aa51b776bb6844b25bce97035b48a9bdb2aea3608687bcdd479d4fa998b5a839ff88558e4a29dff0ed13b55900abb5d439b70793d902ae9ad34587b18c919f6b875c91d14deeb1c373f5e76570d59a6549758f655f1128a54f162dfe8868e1587028e26ad91e528c5ae7ee9335fa58fb59022b5de29d80f0764a9917390d46db899acc6a5b416e25ecc9dccb7153646addcc81cadb5f0078febc7e05d7735aba494f39ef05697bbcc9b47b2ccc79595d75fc13c80678b5e237edce58d731f34c05b1ddcaa649acf2d865bbbc3ceda10508bcdd29d0496744644bf1c3516f6687dfeef5649c7dff90627d642739a59d91a8d1d0c4dc55d74a949e1074427664b467992c9e0f7d3af9d6ea79513e8946ddc0d356bac49878e64e6a95b0a30214214faf2ce317fa622ff3266b32a816e10a18e6d789a5da1f23e67b4f970a68a7bcd9e18825ee274b0483896a40"
-						}
-					],
-					"kernels": [
-						{
-							"features": "Plain",
-							"fee": "7000000",
-							"lock_height": "0",
-							"excess": "000000000000000000000000000000000000000000000000000000000000000000",
-							"excess_sig": "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-						}
-					]
-				}
-			},
-			"amount": "60000000000",
-			"fee": "7000000",
-			"height": "5",
-			"lock_height": "0",
-			"participant_data": [
-				{
-					"id": "0",
-					"public_blind_excess": "033ac2158fa0077f087de60c19d8e431753baa5b63b6e1477f05a2a6e7190d4592",
-					"public_nonce": "031b84c5567b126440995d3ed5aaba0565d71e1834604819ff9c17f5e9d5dd078f",
-					"part_sig": null,
-					"message": null,
-					"message_sig": null
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"slate": {
+				"version_info": {
+					"version": 3,
+					"orig_version": 3,
+					"block_header_version": 2
 				},
-				{
-					"id": "1",
-					"public_blind_excess": "024f9bc78c984c78d6e916d3a00746aa30fa1172124c8dbc0cbddcb7b486719bc7",
-					"public_nonce": "031b84c5567b126440995d3ed5aaba0565d71e1834604819ff9c17f5e9d5dd078f",
-					"part_sig": "8f07ddd5e9f5179cff19486034181ed76505baaad53e5d994064127b56c5841ba9c6dd6185c2b819799700fa1a69201f96cc6dfb9ca205a0ef7c35fb81d57dac",
-					"message": null,
-					"message_sig": null
-				}
-			]
+				"num_participants": 2,
+				"id": "0436430c-2b02-624c-2032-570501212b00",
+				"payment_proof": null,
+				"tx": {
+					"offset": "d202964900000000d302964900000000d402964900000000d502964900000000",
+					"body": {
+						"inputs": [
+							{
+								"features": "Coinbase",
+								"commit": "087df32304c5d4ae8b2af0bc31e700019d722910ef87dd4eec3197b80b207e3045"
+							},
+							{
+								"features": "Coinbase",
+								"commit": "08e1da9e6dc4d6e808a718b2f110a991dd775d65ce5ae408a4e1f002a4961aa9e7"
+							}
+						],
+						"outputs": [
+							{
+								"features": "Plain",
+								"commit": "099b48cfb1f80a2347dc89818449e68e76a3c6817a532a8e9ef2b4a5ccf4363850",
+								"proof": "7ebcd2ed9bf5fb29854033ba3d0e720613bdf7dfacc586d2f6084c1cde0a2b72e955d4ce625916701dc7c347132f40d0f102a34e801d745ee54b49b765d08aae0bb801c60403e57cafade3b4b174e795b633ab9e402b5b1b6e1243fd10bbcf9368a75cb6a6c375c7bdf02da9e03b7f210df45d942e6fba2729cd512a372e6ed91a1b5c9c22831febea843e3f85adcf198f39ac9f7b73b70c60bfb474aa69878ea8d1d32fef30166b59caacaec3fd024de29a90f1587e08d2c36b3d5c560cabf658e212e0a40a4129b3e5c35557058def5551f4eb395759597ba808b3c34eac3bfb9716e4480d7931c5789c538463ec75be0eb807c894047fda6cbcd22682d3c6d3823cb330f090a2099e3510a3706b57d46c95224394d7f1c0a20d99cc314b8f1d9d02668e2e435f62e1194de0be6a1f50f72ed777ed51c8819f527a94918d1aa8df6461e98ed4c2b18210de50fbcf8c3df210bfe326d41f1dc0ad748cb0320ae28401c85ab4f7dcb99d88a052e95dc85b76d22b36cabd60e06ab84bb7e4ddfdab9c9730c8a986583237ed1ecbb323ee8e79b8cadca4b438b7c09531670b471dda6a2eb3e747916c88ce7d9d8e1b7f61660eeb9e5a13c60e4dfe89d1177d81d6f6570fda85158e646a15f1e8b9e977494dc19a339aab2e0e478670d80092d6ba37646e60714ef64eb4a3d37fe15f8f38b59114af34b235489eed3f69b7781c5fe496eb43ffe245c14bd740f745844a38cf0d904347aaa2b64f51add18822dac009d8b63fa3e4c9b1fa72187f9a4acba1ab315daa1b04c9a41f3be846ac420b37990e6c947a16cc9d5c0671b292bf77d7d8b8974d2ad3afae95ba7772c37432840f53a007f31e0195f3abdf100c4477723cc6c6d5da14894a73dfac342833731036487488fdade7b9d556c06f26173b6b67598d3769447ce2828d71dd45ac5af436c6b0"
+							},
+							{
+								"features": "Plain",
+								"commit": "0812276cc788e6870612296d926cba9f0e7b9810670710b5a6e6f1ba006d395774",
+								"proof": "dcff6175390c602bfa92c2ffd1a9b2d84dcc9ea941f6f317bdd0f875244ef23e696fd17c71df79760ce5ce1a96aab1d15dd057358dc835e972febeb86d50ccec0dad7cfe0246d742eb753cf7b88c045d15bc7123f8cf7155647ccf663fca92a83c9a65d0ed756ea7ebffd2cac90c380a102ed9caaa355d175ed0bf58d3ac2f5e909d6c447dfc6b605e04925c2b17c33ebd1908c965a5541ea5d2ed45a0958e6402f89d7a56df1992e036d836e74017e73ccad5cb3a82b8e139e309792a31b15f3ffd72ed033253428c156c2b9799458a25c1da65b719780a22de7fe7f437ae2fccd22cf7ea357ab5aa66a5ef7d71fb0dc64aa0b5761f68278062bb39bb296c787e4cabc5e2a2933a416ce1c9a9696160386449c437e9120f7bb26e5b0e74d1f2e7d5bcd7aafb2a92b87d1548f1f911fb06af7bd6cc13cee29f7c9cb79021aed18186272af0e9d189ec107c81a8a3aeb4782b0d950e4881aa51b776bb6844b25bce97035b48a9bdb2aea3608687bcdd479d4fa998b5a839ff88558e4a29dff0ed13b55900abb5d439b70793d902ae9ad34587b18c919f6b875c91d14deeb1c373f5e76570d59a6549758f655f1128a54f162dfe8868e1587028e26ad91e528c5ae7ee9335fa58fb59022b5de29d80f0764a9917390d46db899acc6a5b416e25ecc9dccb7153646addcc81cadb5f0078febc7e05d7735aba494f39ef05697bbcc9b47b2ccc79595d75fc13c80678b5e237edce58d731f34c05b1ddcaa649acf2d865bbbc3ceda10508bcdd29d0496744644bf1c3516f6687dfeef5649c7dff90627d642739a59d91a8d1d0c4dc55d74a949e1074427664b467992c9e0f7d3af9d6ea79513e8946ddc0d356bac49878e64e6a95b0a30214214faf2ce317fa622ff3266b32a816e10a18e6d789a5da1f23e67b4f970a68a7bcd9e18825ee274b0483896a40"
+							}
+						],
+						"kernels": [
+							{
+								"features": "Plain",
+								"fee": "7000000",
+								"lock_height": "0",
+								"excess": "000000000000000000000000000000000000000000000000000000000000000000",
+								"excess_sig": "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
+							}
+						]
+					}
+				},
+				"amount": "60000000000",
+				"fee": "7000000",
+				"height": "5",
+				"lock_height": "0",
+				"ttl_cutoff_height": null,
+				"participant_data": [
+					{
+						"id": "0",
+						"public_blind_excess": "033ac2158fa0077f087de60c19d8e431753baa5b63b6e1477f05a2a6e7190d4592",
+						"public_nonce": "031b84c5567b126440995d3ed5aaba0565d71e1834604819ff9c17f5e9d5dd078f",
+						"part_sig": null,
+						"message": null,
+						"message_sig": null
+					},
+					{
+						"id": "1",
+						"public_blind_excess": "024f9bc78c984c78d6e916d3a00746aa30fa1172124c8dbc0cbddcb7b486719bc7",
+						"public_nonce": "031b84c5567b126440995d3ed5aaba0565d71e1834604819ff9c17f5e9d5dd078f",
+						"part_sig": "8f07ddd5e9f5179cff19486034181ed76505baaad53e5d994064127b56c5841ba9c6dd6185c2b819799700fa1a69201f96cc6dfb9ca205a0ef7c35fb81d57dac",
+						"message": null,
+						"message_sig": null
+					}
+				]
+			}
 		}
-		]
 	}
 	# "#
 	# ,
@@ -845,10 +889,10 @@ pub trait OwnerRpc: Sync + Send {
 				"fee": "7000000",
 				"height": "5",
 				"id": "0436430c-2b02-624c-2032-570501212b00",
-				"ttl_cutoff_height": null,
-				"payment_proof": null,
 				"lock_height": "0",
+				"ttl_cutoff_height": null,
 				"num_participants": 2,
+				"payment_proof": null,
 				"participant_data": [
 					{
 						"id": "0",
@@ -912,10 +956,11 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# , false, 5, true, true, false, false);
+	# , true, 5, true, true, false, false);
 	```
 	 */
-	fn finalize_tx(&self, slate: VersionedSlate) -> Result<VersionedSlate, ErrorKind>;
+	fn finalize_tx(&self, token: Token, slate: VersionedSlate)
+		-> Result<VersionedSlate, ErrorKind>;
 
 	/**
 	Networked version of [Owner::post_tx](struct.Owner.html#method.post_tx).
@@ -927,8 +972,9 @@ pub trait OwnerRpc: Sync + Send {
 		"jsonrpc": "2.0",
 		"id": 1,
 		"method": "post_tx",
-		"params": [
-		{
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"tx": {
 			"offset": "d202964900000000d302964900000000d402964900000000d502964900000000",
 			"body": {
 				"inputs": [
@@ -964,8 +1010,8 @@ pub trait OwnerRpc: Sync + Send {
 				]
 			}
 		},
-		false
-		]
+		"fluff": false
+		}
 	}
 	# "#
 	# ,
@@ -978,11 +1024,11 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# , false, 5, true, true, true, false);
+	# , true, 5, true, true, true, false);
 	```
 	 */
 
-	fn post_tx(&self, tx: TransactionV3, fluff: bool) -> Result<(), ErrorKind>;
+	fn post_tx(&self, token: Token, tx: TransactionV3, fluff: bool) -> Result<(), ErrorKind>;
 
 	/**
 	Networked version of [Owner::cancel_tx](struct.Owner.html#method.cancel_tx).
@@ -994,7 +1040,11 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "cancel_tx",
-		"params": [null, "0436430c-2b02-624c-2032-570501212b00"],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"tx_id": null,
+			"tx_slate_id": "0436430c-2b02-624c-2032-570501212b00"
+		},
 		"id": 1
 	}
 	# "#
@@ -1008,10 +1058,15 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# , false, 5, true, true, false, false);
+	# , true, 5, true, true, false, false);
 	```
 	 */
-	fn cancel_tx(&self, tx_id: Option<u32>, tx_slate_id: Option<Uuid>) -> Result<(), ErrorKind>;
+	fn cancel_tx(
+		&self,
+		token: Token,
+		tx_id: Option<u32>,
+		tx_slate_id: Option<Uuid>,
+	) -> Result<(), ErrorKind>;
 
 	/**
 	Networked version of [Owner::get_stored_tx](struct.Owner.html#method.get_stored_tx).
@@ -1023,8 +1078,9 @@ pub trait OwnerRpc: Sync + Send {
 		"jsonrpc": "2.0",
 		"method": "get_stored_tx",
 		"id": 1,
-		"params": [
-			{
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"tx": {
 				"amount_credited": "59993000000",
 				"amount_debited": "120000000000",
 				"confirmation_ts": "2019-01-15T16:01:26Z",
@@ -1057,7 +1113,7 @@ pub trait OwnerRpc: Sync + Send {
 				"kernel_excess": null,
 				"kernel_lookup_min_height": null
 			}
-		]
+		}
 	}
 	# "#
 	# ,
@@ -1105,10 +1161,14 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# , false, 5, true, true, false, false);
+	# , true, 5, true, true, false, false);
 	```
 	 */
-	fn get_stored_tx(&self, tx: &TxLogEntry) -> Result<Option<TransactionV3>, ErrorKind>;
+	fn get_stored_tx(
+		&self,
+		token: Token,
+		tx: &TxLogEntry,
+	) -> Result<Option<TransactionV3>, ErrorKind>;
 
 	/**
 	Networked version of [Owner::verify_slate_messages](struct.Owner.html#method.verify_slate_messages).
@@ -1120,14 +1180,15 @@ pub trait OwnerRpc: Sync + Send {
 		"jsonrpc": "2.0",
 		"method": "verify_slate_messages",
 		"id": 1,
-		"params": [ {
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"slate": {
 				"amount": "6000000000",
 				"fee": "8000000",
 				"height": "4",
 				"id": "0436430c-2b02-624c-2032-570501212b00",
 				"lock_height": "4",
 				"ttl_cutoff_height": null,
-				"payment_proof": null,
 				"num_participants": 2,
 				"participant_data": [
 				{
@@ -1164,7 +1225,8 @@ pub trait OwnerRpc: Sync + Send {
 						}
 						]
 					},
-					"offset": "d202964900000000d302964900000000d402964900000000d502964900000000"
+					"offset": "d202964900000000d302964900000000d402964900000000d502964900000000",
+					"payment_proof": null
 				},
 				"version_info": {
 					"orig_version": 3,
@@ -1172,7 +1234,7 @@ pub trait OwnerRpc: Sync + Send {
 					"block_header_version": 2
 				}
 			}
-		]
+		}
 	}
 	# "#
 	# ,
@@ -1185,10 +1247,10 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# ,false, 0 ,false, false, false, false);
+	# ,true, 0 ,false, false, false, false);
 	```
 	*/
-	fn verify_slate_messages(&self, slate: VersionedSlate) -> Result<(), ErrorKind>;
+	fn verify_slate_messages(&self, token: Token, slate: VersionedSlate) -> Result<(), ErrorKind>;
 
 	/**
 	Networked version of [Owner::scan](struct.Owner.html#method.scan).
@@ -1200,7 +1262,11 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "scan",
-		"params": [null, false],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"start_height": 1,
+			"delete_unconfirmed": false
+		},
 		"id": 1
 	}
 	# "#
@@ -1214,14 +1280,18 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# , false, 1, false, false, false, false);
+	# , true, 1, false, false, false, false);
 	```
 	 */
-	fn scan(&self, start_height: Option<u64>, delete_unconfirmed: bool) -> Result<(), ErrorKind>;
+	fn scan(
+		&self,
+		token: Token,
+		start_height: Option<u64>,
+		delete_unconfirmed: bool,
+	) -> Result<(), ErrorKind>;
 
 	/**
 	Networked version of [Owner::node_height](struct.Owner.html#method.node_height).
-
 
 	```
 	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
@@ -1229,7 +1299,9 @@ pub trait OwnerRpc: Sync + Send {
 	{
 		"jsonrpc": "2.0",
 		"method": "node_height",
-		"params": [],
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000"
+		},
 		"id": 1
 	}
 	# "#
@@ -1247,121 +1319,1020 @@ pub trait OwnerRpc: Sync + Send {
 		}
 	}
 	# "#
-	# , false, 5, false, false, false, false);
+	# , true, 5, false, false, false, false);
 	```
 	 */
-	fn node_height(&self) -> Result<NodeHeightResult, ErrorKind>;
+	fn node_height(&self, token: Token) -> Result<NodeHeightResult, ErrorKind>;
+
+	/**
+		Initializes the secure JSON-RPC API. This function must be called and a shared key
+		established before any other OwnerAPI JSON-RPC function can be called.
+
+		The shared key will be derived using ECDH with the provided public key on the secp256k1 curve. This
+		function will return its public key used in the derivation, which the caller should multiply by its
+		private key to derive the shared key.
+
+		Once the key is established, all further requests and responses are encrypted and decrypted with the
+		following parameters:
+		* AES-256 in GCM mode with 128-bit tags and 96 bit nonces
+		* 12 byte nonce which must be included in each request/response to use on the decrypting side
+		* Empty vector for additional data
+		* Suffix length = AES-256 GCM mode tag length = 16 bytes
+		*
+
+		Fully-formed JSON-RPC requests (as documented) should be encrypted using these parameters, encoded
+		into base64 and included with the one-time nonce in a request for the `encrypted_request_v3` method
+		as follows:
+
+		```
+		# let s = r#"
+		{
+			 "jsonrpc": "2.0",
+			 "method": "encrypted_request_v3",
+			 "id": "1",
+			 "params": {
+					"nonce": "ef32...",
+					"body_enc": "e0bcd..."
+			 }
+		}
+		# "#;
+		```
+
+		With a typical response being:
+
+		```
+		# let s = r#"{
+		{
+			 "jsonrpc": "2.0",
+			 "method": "encrypted_response_v3",
+			 "id": "1",
+			 "Ok": {
+					"nonce": "340b...",
+					"body_enc": "3f09c..."
+			 }
+		}
+		# }"#;
+		```
+
+	*/
+
+	fn init_secure_api(&self, ecdh_pubkey: ECDHPubkey) -> Result<ECDHPubkey, ErrorKind>;
+
+	/**
+	Networked version of [Owner::get_top_level_directory](struct.Owner.html#method.get_top_level_directory).
+
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "get_top_level_directory",
+		"params": {
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": "/doctest/dir"
+		}
+	}
+	# "#
+	# , true, 5, false, false, false, false);
+	```
+	*/
+
+	fn get_top_level_directory(&self) -> Result<String, ErrorKind>;
+
+	/**
+	Networked version of [Owner::set_top_level_directory](struct.Owner.html#method.set_top_level_directory).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "set_top_level_directory",
+		"params": {
+			"dir": "/home/wallet_user/my_wallet_dir"
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 5, false, false, false, false);
+	```
+	*/
+
+	fn set_top_level_directory(&self, dir: String) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::create_config](struct.Owner.html#method.create_config).
+
+	Both the `wallet_config` and `logging_config` parameters can be `null`, the examples
+	below are for illustration. Note that the values provided for `log_file_path` and `data_file_dir`
+	will be ignored and replaced with the actual values based on the value of `get_top_level_directory`
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "create_config",
+		"params": {
+			"chain_type": "Mainnet",
+			"wallet_config": {
+				"chain_type": null,
+				"api_listen_interface": "127.0.0.1",
+				"api_listen_port": 3415,
+				"owner_api_listen_port": 3420,
+				"api_secret_path": null,
+				"node_api_secret_path": null,
+				"check_node_api_http_addr": "http://127.0.0.1:3413",
+				"owner_api_include_foreign": false,
+				"data_file_dir": "/path/to/data/file/dir",
+				"no_commit_cache": null,
+				"tls_certificate_file": null,
+				"tls_certificate_key": null,
+				"dark_background_color_scheme": null,
+				"keybase_notify_ttl": null
+			},
+			"logging_config": {
+				"log_to_stdout": false,
+				"stdout_log_level": "Info",
+				"log_to_file": true,
+				"file_log_level": "Debug",
+				"log_file_path": "/path/to/log/file",
+				"log_file_append": true,
+				"log_max_size": null,
+				"log_max_files": null,
+				"tui_running": null
+			},
+			"tor_config" : {
+				"use_tor_listener": true,
+				"socks_proxy_addr": "127.0.0.1:9050",
+				"send_config_dir": "."
+			}
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 5, false, false, false, false);
+	```
+	*/
+	fn create_config(
+		&self,
+		chain_type: global::ChainTypes,
+		wallet_config: Option<WalletConfig>,
+		logging_config: Option<LoggingConfig>,
+		tor_config: Option<TorConfig>,
+	) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::create_wallet](struct.Owner.html#method.create_wallet).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "create_wallet",
+		"params": {
+			"name": null,
+			"mnemonic": null,
+			"mnemonic_length": 32,
+			"password": "my_secret_password"
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn create_wallet(
+		&self,
+		name: Option<String>,
+		mnemonic: Option<String>,
+		mnemonic_length: u32,
+		password: String,
+	) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::open_wallet](struct.Owner.html#method.open_wallet).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "open_wallet",
+		"params": {
+			"name": null,
+			"password": "my_secret_password"
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": "d096b3cb75986b3b13f80b8f5243a9edf0af4c74ac37578c5a12cfb5b59b1868"
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn open_wallet(&self, name: Option<String>, password: String) -> Result<Token, ErrorKind>;
+
+	/**
+	Networked version of [Owner::close_wallet](struct.Owner.html#method.close_wallet).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "close_wallet",
+		"params": {
+			"name": null
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn close_wallet(&self, name: Option<String>) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::get_mnemonic](struct.Owner.html#method.get_mnemonic).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "get_mnemonic",
+		"params": {
+			"name": null,
+			"password": ""
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": "fat twenty mean degree forget shell check candy immense awful flame next during february bulb bike sun wink theory day kiwi embrace peace lunch"
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn get_mnemonic(&self, name: Option<String>, password: String) -> Result<String, ErrorKind>;
+
+	/**
+	Networked version of [Owner::change_password](struct.Owner.html#method.change_password).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "change_password",
+		"params": {
+			"name": null,
+			"old": "",
+			"new": "new_password"
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+	fn change_password(
+		&self,
+		name: Option<String>,
+		old: String,
+		new: String,
+	) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::delete_wallet](struct.Owner.html#method.delete_wallet).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "delete_wallet",
+		"params": {
+			"name": null
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+	fn delete_wallet(&self, name: Option<String>) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::start_updated](struct.Owner.html#method.start_updater).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "start_updater",
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"frequency": 30000
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn start_updater(&self, token: Token, frequency: u32) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::stop_updater](struct.Owner.html#method.stop_updater).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "stop_updater",
+		"params": null,
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+	fn stop_updater(&self) -> Result<(), ErrorKind>;
+
+	/**
+	Networked version of [Owner::get_updater_messages](struct.Owner.html#method.get_updater_messages).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "get_updater_messages",
+		"params": {
+			"count": 1
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": []
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn get_updater_messages(&self, count: u32) -> Result<Vec<StatusMessage>, ErrorKind>;
+
+	/**
+	Networked version of [Owner::get_public_proof_address](struct.Owner.html#method.get_public_proof_address).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "get_public_proof_address",
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"derivation_index": 0
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": "32cdd63928854f8b2628b1dce4626ddcdf35d56cb7cfdf7d64cca5822b78d4d3"
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn get_public_proof_address(
+		&self,
+		token: Token,
+		derivation_index: u32,
+	) -> Result<PubAddress, ErrorKind>;
+
+	/**
+	Networked version of [Owner::proof_address_from_onion_v3](struct.Owner.html#method.proof_address_from_onion_v3).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "proof_address_from_onion_v3",
+		"params": {
+			"address_v3": "2a6at2obto3uvkpkitqp4wxcg6u36qf534eucbskqciturczzc5suyid"
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": "d03c09e9c19bb74aa9ea44e0fe5ae237a9bf40bddf0941064a80913a4459c8bb"
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+
+	fn proof_address_from_onion_v3(&self, address_v3: String) -> Result<PubAddress, ErrorKind>;
+
+	/**
+	Networked version of [Owner::retrieve_payment_proof](struct.Owner.html#method.retrieve_payment_proof).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "retrieve_payment_proof",
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"refresh_from_node": true,
+			"tx_id": null,
+			"tx_slate_id": "0436430c-2b02-624c-2032-570501212b00"
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": {
+				"amount": "60000000000",
+				"excess": "09bac6083b05a32a9d9b37710c70dd0a1ef9329fde0848558976b6f1b81d80ceed",
+				"recipient_address": "pa7wkkdgs5bkteha7lykl7ff2wztgdrxxo442xdcq2lnaphe5aidd4id",
+				"recipient_sig": "42b6f2bbcee432185993867d1a338e260454ead536bf728f4dcc8f535508715e92a0695486ba3c9945d8ecb2cf7f703a955780253b12d0048f02d318c8f08702",
+				"sender_address": "glg5mojiqvhywjriwhooiytn3tptlvlmw7h567lezssyek3y2tjzznad",
+				"sender_sig": "5e3f5596852e83f6db7152fc51c41b4ed8742eb8045fa85a6965c52d09fcb46ba67d4f86660c9f3dc55ab84faea79d11c3831aa77934f7e90695e63d523f8604"
+			}
+		}
+	}
+	# "#
+	# , true, 5, true, true, true, true);
+	```
+	*/
+
+	fn retrieve_payment_proof(
+		&self,
+		token: Token,
+		refresh_from_node: bool,
+		tx_id: Option<u32>,
+		tx_slate_id: Option<Uuid>,
+	) -> Result<PaymentProof, ErrorKind>;
+
+	/**
+	Networked version of [Owner::verify_payment_proof](struct.Owner.html#method.verify_payment_proof).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "verify_payment_proof",
+		"params": {
+			"token": "d202964900000000d302964900000000d402964900000000d502964900000000",
+			"proof": {
+				"amount": "60000000000",
+				"excess": "09bac6083b05a32a9d9b37710c70dd0a1ef9329fde0848558976b6f1b81d80ceed",
+				"recipient_address": "pa7wkkdgs5bkteha7lykl7ff2wztgdrxxo442xdcq2lnaphe5aidd4id",
+				"recipient_sig": "42b6f2bbcee432185993867d1a338e260454ead536bf728f4dcc8f535508715e92a0695486ba3c9945d8ecb2cf7f703a955780253b12d0048f02d318c8f08702",
+				"sender_address": "glg5mojiqvhywjriwhooiytn3tptlvlmw7h567lezssyek3y2tjzznad",
+				"sender_sig": "5e3f5596852e83f6db7152fc51c41b4ed8742eb8045fa85a6965c52d09fcb46ba67d4f86660c9f3dc55ab84faea79d11c3831aa77934f7e90695e63d523f8604"
+			}
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": [
+				true,
+				false
+			]
+		}
+	}
+	# "#
+	# , true, 5, true, true, true, true);
+	```
+	*/
+
+	fn verify_payment_proof(
+		&self,
+		token: Token,
+		proof: PaymentProof,
+	) -> Result<(bool, bool), ErrorKind>;
+
+	/**
+	Networked version of [Owner::set_tor_config](struct.Owner.html#method.set_tor_config).
+	```
+	# grin_wallet_api::doctest_helper_json_rpc_owner_assert_response!(
+	# r#"
+	{
+		"jsonrpc": "2.0",
+		"method": "set_tor_config",
+		"params": {
+			"tor_config": {
+				"use_tor_listener": true,
+				"socks_proxy_addr": "127.0.0.1:59050",
+				"send_config_dir": "."
+			}
+		},
+		"id": 1
+	}
+	# "#
+	# ,
+	# r#"
+	{
+		"id": 1,
+		"jsonrpc": "2.0",
+		"result": {
+			"Ok": null
+		}
+	}
+	# "#
+	# , true, 0, false, false, false, false);
+	```
+	*/
+	fn set_tor_config(&self, tor_config: Option<TorConfig>) -> Result<(), ErrorKind>;
 }
 
-impl<'a, L, C, K> OwnerRpc for Owner<L, C, K>
+impl<L, C, K> OwnerRpc for Owner<L, C, K>
 where
 	L: WalletLCProvider<'static, C, K>,
 	C: NodeClient + 'static,
 	K: Keychain + 'static,
 {
-	fn accounts(&self) -> Result<Vec<AcctPathMapping>, ErrorKind> {
-		Owner::accounts(self, None).map_err(|e| e.kind())
+	fn accounts(&self, token: Token) -> Result<Vec<AcctPathMapping>, ErrorKind> {
+		Owner::accounts(self, (&token.keychain_mask).as_ref()).map_err(|e| e.kind())
 	}
 
-	fn create_account_path(&self, label: &String) -> Result<Identifier, ErrorKind> {
-		Owner::create_account_path(self, None, label).map_err(|e| e.kind())
+	fn create_account_path(&self, token: Token, label: &String) -> Result<Identifier, ErrorKind> {
+		Owner::create_account_path(self, (&token.keychain_mask).as_ref(), label)
+			.map_err(|e| e.kind())
 	}
 
-	fn set_active_account(&self, label: &String) -> Result<(), ErrorKind> {
-		Owner::set_active_account(self, None, label).map_err(|e| e.kind())
+	fn set_active_account(&self, token: Token, label: &String) -> Result<(), ErrorKind> {
+		Owner::set_active_account(self, (&token.keychain_mask).as_ref(), label)
+			.map_err(|e| e.kind())
 	}
 
 	fn retrieve_outputs(
 		&self,
+		token: Token,
 		include_spent: bool,
 		refresh_from_node: bool,
 		tx_id: Option<u32>,
 	) -> Result<(bool, Vec<OutputCommitMapping>), ErrorKind> {
-		Owner::retrieve_outputs(self, None, include_spent, refresh_from_node, tx_id)
-			.map_err(|e| e.kind())
+		Owner::retrieve_outputs(
+			self,
+			(&token.keychain_mask).as_ref(),
+			include_spent,
+			refresh_from_node,
+			tx_id,
+		)
+		.map_err(|e| e.kind())
 	}
 
 	fn retrieve_txs(
 		&self,
+		token: Token,
 		refresh_from_node: bool,
 		tx_id: Option<u32>,
 		tx_slate_id: Option<Uuid>,
 	) -> Result<(bool, Vec<TxLogEntry>), ErrorKind> {
-		Owner::retrieve_txs(self, None, refresh_from_node, tx_id, tx_slate_id).map_err(|e| e.kind())
+		Owner::retrieve_txs(
+			self,
+			(&token.keychain_mask).as_ref(),
+			refresh_from_node,
+			tx_id,
+			tx_slate_id,
+		)
+		.map_err(|e| e.kind())
 	}
 
 	fn retrieve_summary_info(
 		&self,
+		token: Token,
 		refresh_from_node: bool,
 		minimum_confirmations: u64,
 	) -> Result<(bool, WalletInfo), ErrorKind> {
-		Owner::retrieve_summary_info(self, None, refresh_from_node, minimum_confirmations)
-			.map_err(|e| e.kind())
+		Owner::retrieve_summary_info(
+			self,
+			(&token.keychain_mask).as_ref(),
+			refresh_from_node,
+			minimum_confirmations,
+		)
+		.map_err(|e| e.kind())
 	}
 
-	fn init_send_tx(&self, args: InitTxArgs) -> Result<VersionedSlate, ErrorKind> {
-		let slate = Owner::init_send_tx(self, None, args).map_err(|e| e.kind())?;
+	fn init_send_tx(&self, token: Token, args: InitTxArgs) -> Result<VersionedSlate, ErrorKind> {
+		let slate = Owner::init_send_tx(self, (&token.keychain_mask).as_ref(), args)
+			.map_err(|e| e.kind())?;
 		let version = SlateVersion::V3;
 		Ok(VersionedSlate::into_version(slate, version))
 	}
 
-	fn issue_invoice_tx(&self, args: IssueInvoiceTxArgs) -> Result<VersionedSlate, ErrorKind> {
-		let slate = Owner::issue_invoice_tx(self, None, args).map_err(|e| e.kind())?;
+	fn issue_invoice_tx(
+		&self,
+		token: Token,
+		args: IssueInvoiceTxArgs,
+	) -> Result<VersionedSlate, ErrorKind> {
+		let slate = Owner::issue_invoice_tx(self, (&token.keychain_mask).as_ref(), args)
+			.map_err(|e| e.kind())?;
 		let version = SlateVersion::V3;
 		Ok(VersionedSlate::into_version(slate, version))
 	}
 
 	fn process_invoice_tx(
 		&self,
+		token: Token,
 		in_slate: VersionedSlate,
 		args: InitTxArgs,
 	) -> Result<VersionedSlate, ErrorKind> {
-		let out_slate = Owner::process_invoice_tx(self, None, &Slate::from(in_slate), args)
-			.map_err(|e| e.kind())?;
+		let out_slate = Owner::process_invoice_tx(
+			self,
+			(&token.keychain_mask).as_ref(),
+			&Slate::from(in_slate),
+			args,
+		)
+		.map_err(|e| e.kind())?;
 		let version = SlateVersion::V3;
 		Ok(VersionedSlate::into_version(out_slate, version))
 	}
 
-	fn finalize_tx(&self, in_slate: VersionedSlate) -> Result<VersionedSlate, ErrorKind> {
-		let out_slate =
-			Owner::finalize_tx(self, None, &Slate::from(in_slate)).map_err(|e| e.kind())?;
+	fn finalize_tx(
+		&self,
+		token: Token,
+		in_slate: VersionedSlate,
+	) -> Result<VersionedSlate, ErrorKind> {
+		let out_slate = Owner::finalize_tx(
+			self,
+			(&token.keychain_mask).as_ref(),
+			&Slate::from(in_slate),
+		)
+		.map_err(|e| e.kind())?;
 		let version = SlateVersion::V3;
 		Ok(VersionedSlate::into_version(out_slate, version))
 	}
 
 	fn tx_lock_outputs(
 		&self,
-		slate: VersionedSlate,
+		token: Token,
+		in_slate: VersionedSlate,
 		participant_id: usize,
 	) -> Result<(), ErrorKind> {
-		Owner::tx_lock_outputs(self, None, &Slate::from(slate), participant_id)
+		Owner::tx_lock_outputs(
+			self,
+			(&token.keychain_mask).as_ref(),
+			&Slate::from(in_slate),
+			participant_id,
+		)
+		.map_err(|e| e.kind())
+	}
+
+	fn cancel_tx(
+		&self,
+		token: Token,
+		tx_id: Option<u32>,
+		tx_slate_id: Option<Uuid>,
+	) -> Result<(), ErrorKind> {
+		Owner::cancel_tx(self, (&token.keychain_mask).as_ref(), tx_id, tx_slate_id)
 			.map_err(|e| e.kind())
 	}
 
-	fn cancel_tx(&self, tx_id: Option<u32>, tx_slate_id: Option<Uuid>) -> Result<(), ErrorKind> {
-		Owner::cancel_tx(self, None, tx_id, tx_slate_id).map_err(|e| e.kind())
-	}
-
-	fn get_stored_tx(&self, tx: &TxLogEntry) -> Result<Option<TransactionV3>, ErrorKind> {
-		Owner::get_stored_tx(self, None, tx)
+	fn get_stored_tx(
+		&self,
+		token: Token,
+		tx: &TxLogEntry,
+	) -> Result<Option<TransactionV3>, ErrorKind> {
+		Owner::get_stored_tx(self, (&token.keychain_mask).as_ref(), tx)
 			.map(|x| x.map(TransactionV3::from))
 			.map_err(|e| e.kind())
 	}
 
-	fn post_tx(&self, tx: TransactionV3, fluff: bool) -> Result<(), ErrorKind> {
-		Owner::post_tx(self, None, &Transaction::from(tx), fluff).map_err(|e| e.kind())
+	fn post_tx(&self, token: Token, tx: TransactionV3, fluff: bool) -> Result<(), ErrorKind> {
+		Owner::post_tx(
+			self,
+			(&token.keychain_mask).as_ref(),
+			&Transaction::from(tx),
+			fluff,
+		)
+		.map_err(|e| e.kind())
 	}
 
-	fn verify_slate_messages(&self, slate: VersionedSlate) -> Result<(), ErrorKind> {
-		Owner::verify_slate_messages(self, None, &Slate::from(slate)).map_err(|e| e.kind())
+	fn verify_slate_messages(&self, token: Token, slate: VersionedSlate) -> Result<(), ErrorKind> {
+		Owner::verify_slate_messages(self, (&token.keychain_mask).as_ref(), &Slate::from(slate))
+			.map_err(|e| e.kind())
 	}
 
-	fn scan(&self, start_height: Option<u64>, delete_unconfirmed: bool) -> Result<(), ErrorKind> {
-		Owner::scan(self, None, start_height, delete_unconfirmed).map_err(|e| e.kind())
+	fn scan(
+		&self,
+		token: Token,
+		start_height: Option<u64>,
+		delete_unconfirmed: bool,
+	) -> Result<(), ErrorKind> {
+		Owner::scan(
+			self,
+			(&token.keychain_mask).as_ref(),
+			start_height,
+			delete_unconfirmed,
+		)
+		.map_err(|e| e.kind())
 	}
 
-	fn node_height(&self) -> Result<NodeHeightResult, ErrorKind> {
-		Owner::node_height(self, None).map_err(|e| e.kind())
+	fn node_height(&self, token: Token) -> Result<NodeHeightResult, ErrorKind> {
+		Owner::node_height(self, (&token.keychain_mask).as_ref()).map_err(|e| e.kind())
+	}
+
+	fn init_secure_api(&self, ecdh_pubkey: ECDHPubkey) -> Result<ECDHPubkey, ErrorKind> {
+		let secp_inst = static_secp_instance();
+		let secp = secp_inst.lock();
+		let sec_key = SecretKey::new(&secp, &mut thread_rng());
+
+		let mut shared_pubkey = ecdh_pubkey.ecdh_pubkey;
+		shared_pubkey
+			.mul_assign(&secp, &sec_key)
+			.map_err(ErrorKind::Secp)?;
+
+		let x_coord = shared_pubkey.serialize_vec(&secp, true);
+		let shared_key = SecretKey::from_slice(&secp, &x_coord[1..]).map_err(ErrorKind::Secp)?;
+		{
+			let mut s = self.shared_key.lock();
+			*s = Some(shared_key);
+		}
+
+		let pub_key = PublicKey::from_secret_key(&secp, &sec_key).map_err(ErrorKind::Secp)?;
+
+		Ok(ECDHPubkey {
+			ecdh_pubkey: pub_key,
+		})
+	}
+
+	fn get_top_level_directory(&self) -> Result<String, ErrorKind> {
+		Owner::get_top_level_directory(self).map_err(|e| e.kind())
+	}
+
+	fn set_top_level_directory(&self, dir: String) -> Result<(), ErrorKind> {
+		Owner::set_top_level_directory(self, &dir).map_err(|e| e.kind())
+	}
+
+	fn create_config(
+		&self,
+		chain_type: global::ChainTypes,
+		wallet_config: Option<WalletConfig>,
+		logging_config: Option<LoggingConfig>,
+		tor_config: Option<TorConfig>,
+	) -> Result<(), ErrorKind> {
+		Owner::create_config(self, &chain_type, wallet_config, logging_config, tor_config)
+			.map_err(|e| e.kind())
+	}
+
+	fn create_wallet(
+		&self,
+		name: Option<String>,
+		mnemonic: Option<String>,
+		mnemonic_length: u32,
+		password: String,
+	) -> Result<(), ErrorKind> {
+		let n = name.as_ref().map(|s| s.as_str());
+		let m = match mnemonic {
+			Some(s) => Some(ZeroingString::from(s)),
+			None => None,
+		};
+		Owner::create_wallet(self, n, m, mnemonic_length, ZeroingString::from(password))
+			.map_err(|e| e.kind())
+	}
+
+	fn open_wallet(&self, name: Option<String>, password: String) -> Result<Token, ErrorKind> {
+		let n = name.as_ref().map(|s| s.as_str());
+		let sec_key = Owner::open_wallet(self, n, ZeroingString::from(password), true)
+			.map_err(|e| e.kind())?;
+		Ok(Token {
+			keychain_mask: sec_key,
+		})
+	}
+
+	fn close_wallet(&self, name: Option<String>) -> Result<(), ErrorKind> {
+		let n = name.as_ref().map(|s| s.as_str());
+		Owner::close_wallet(self, n).map_err(|e| e.kind())
+	}
+
+	fn get_mnemonic(&self, name: Option<String>, password: String) -> Result<String, ErrorKind> {
+		let n = name.as_ref().map(|s| s.as_str());
+		let res =
+			Owner::get_mnemonic(self, n, ZeroingString::from(password)).map_err(|e| e.kind())?;
+		Ok((&*res).to_string())
+	}
+
+	fn change_password(
+		&self,
+		name: Option<String>,
+		old: String,
+		new: String,
+	) -> Result<(), ErrorKind> {
+		let n = name.as_ref().map(|s| s.as_str());
+		Owner::change_password(self, n, ZeroingString::from(old), ZeroingString::from(new))
+			.map_err(|e| e.kind())
+	}
+
+	fn delete_wallet(&self, name: Option<String>) -> Result<(), ErrorKind> {
+		let n = name.as_ref().map(|s| s.as_str());
+		Owner::delete_wallet(self, n).map_err(|e| e.kind())
+	}
+
+	fn start_updater(&self, token: Token, frequency: u32) -> Result<(), ErrorKind> {
+		Owner::start_updater(
+			self,
+			(&token.keychain_mask).as_ref(),
+			Duration::from_millis(frequency as u64),
+		)
+		.map_err(|e| e.kind())
+	}
+
+	fn stop_updater(&self) -> Result<(), ErrorKind> {
+		Owner::stop_updater(self).map_err(|e| e.kind())
+	}
+
+	fn get_updater_messages(&self, count: u32) -> Result<Vec<StatusMessage>, ErrorKind> {
+		Owner::get_updater_messages(self, count as usize).map_err(|e| e.kind())
+	}
+
+	fn get_public_proof_address(
+		&self,
+		token: Token,
+		derivation_index: u32,
+	) -> Result<PubAddress, ErrorKind> {
+		let address = Owner::get_public_proof_address(
+			self,
+			(&token.keychain_mask).as_ref(),
+			derivation_index,
+		)
+		.map_err(|e| e.kind())?;
+		Ok(PubAddress { address })
+	}
+
+	fn retrieve_payment_proof(
+		&self,
+		token: Token,
+		refresh_from_node: bool,
+		tx_id: Option<u32>,
+		tx_slate_id: Option<Uuid>,
+	) -> Result<PaymentProof, ErrorKind> {
+		Owner::retrieve_payment_proof(
+			self,
+			(&token.keychain_mask).as_ref(),
+			refresh_from_node,
+			tx_id,
+			tx_slate_id,
+		)
+		.map_err(|e| e.kind())
+	}
+
+	fn verify_payment_proof(
+		&self,
+		token: Token,
+		proof: PaymentProof,
+	) -> Result<(bool, bool), ErrorKind> {
+		Owner::verify_payment_proof(self, (&token.keychain_mask).as_ref(), &proof)
+			.map_err(|e| e.kind())
+	}
+
+	fn proof_address_from_onion_v3(&self, address_v3: String) -> Result<PubAddress, ErrorKind> {
+		let address =
+			Owner::proof_address_from_onion_v3(self, &address_v3).map_err(|e| e.kind())?;
+		Ok(PubAddress { address })
+	}
+
+	fn set_tor_config(&self, tor_config: Option<TorConfig>) -> Result<(), ErrorKind> {
+		Owner::set_tor_config(self, tor_config);
+		Ok(())
 	}
 }
 
@@ -1382,12 +2353,10 @@ pub fn run_doctest_owner(
 	use grin_wallet_libwallet::{api_impl, WalletInst};
 	use grin_wallet_util::grin_keychain::ExtKeychain;
 
-	use crate::core::global;
 	use crate::core::global::ChainTypes;
 	use grin_wallet_util::grin_util as util;
 
-	use std::fs;
-	use std::thread;
+	use std::{fs, thread};
 
 	util::init_test_logger();
 	let _ = fs::remove_dir_all(test_dir);
@@ -1572,13 +2541,8 @@ pub fn run_doctest_owner(
 
 	let mut api_owner = Owner::new(wallet1, None);
 	api_owner.doctest_mode = true;
-	let res = if use_token {
-		let owner_api = &api_owner as &dyn OwnerRpcS;
-		owner_api.handle_request(request).as_option()
-	} else {
-		let owner_api = &api_owner as &dyn OwnerRpc;
-		owner_api.handle_request(request).as_option()
-	};
+	let owner_api = &api_owner as &dyn OwnerRpc;
+	let res = owner_api.handle_request(request).as_option();
 	let _ = fs::remove_dir_all(test_dir);
 	Ok(res)
 }
